Cleaned up the project renamed some components and some views, made a folder for my lists. Removoed 2 out of 3 modals and made one of them generic for reuse in different places.
