+++ src/components/Modal.vue
@@ -37,7 +37,7 @@
 
 <script>
 export default{
-    name: 'Bookingsmodal',
+    name: 'Modal',
 
     methods: {
         close() {
